requirements.txt
pytest==8.3.3
pytest-html==4.1.1
appium-python-client==4.1.0
selenium==4.27.1
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
faker==30.8.2
jsonschema==4.23.0
allure-pytest==2.13.5
webdriver-manager==4.0.2
python-dotenv==1.0.1